move tests to right place
